Create Pros Bot Test
This was a test for a basic program in pros and to confirm that the debug stream could be used. I chose to use PROS from now on because I could not find a viable way to run the data that was previously recorded during the autonomous method. I am hoping that PROS will be able to run off recorded data.